fix long

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int c = 0;
+	unsigned long int c = 0;
 
 	if (n < index)
 		return (-1);
